Switched from index-based to iterator-based access in the choice enumeration method for greater efficiency
Access to an object is now incremental with the loop over the selection mask instead of being index-based (which required to start over at every iteration). Although it may not have a significant impact on vectors, this change makes the algorithm more efficient overall, especially when using lists or list-like containers.